manager/requirements_.txt
Django==3.2
django-cors-headers==3.14.0
django-filter==22.1
django-mysql==4.8.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
# mysqlclient>=2.1.1<3.1
mysqlclient==2.2.4
PyJWT==2.6.0
Pillow==9.0.0
boto3==1.20.46
botocore==1.23.46
django-storages==1.12.3
drf-spectacular>=0.15.1,<0.16